requirements.txt
adium-theme-ubuntu==0.3.4
dnspython==1.12.0
ecdsa==0.13
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
mysql-connector-python-rf==2.2.2
numpy==1.13.0
pandas==0.20.2
paramiko==1.16.0
pycrypto==2.6.1
pymongo==2.8
python-dateutil==2.6.0
pytz==2017.2
Shapely==1.5.17.post1
six==1.10.0
unity-lens-photos==1.0
Werkzeug==0.10
